NodeGraphTest : Move GraphGadget tests into GraphGadgetTest.
