+++ python/GafferUITest/__init__.py
@@ -104,6 +104,7 @@ from DocumentationTest import DocumentationTest
 from LazyMethodTest import LazyMethodTest
 from ReferenceUITest import ReferenceUITest
 from CompoundDataPlugValueWidgetTest import CompoundDataPlugValueWidgetTest
+from GraphGadgetTest import GraphGadgetTest
 
 if __name__ == "__main__":
 	unittest.main()
